refactor(familiars): separate roster and detail sections
