+++ src/components/familiars-view-sections.tsx
@@ -1,0 +1,691 @@
+"use client";
+
+import Link from "next/link";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { Icon } from "@/lib/icon";
+import { usePausablePoll } from "@/lib/use-pausable-poll";
+import { useFocusTrap } from "@/lib/use-focus-trap";
+import { Tabs } from "@/components/ui/tabs";
+// Shared relative-time formatter, imported as `age` so the call sites read the
+// same — standardizes this surface on the app-wide "2m ago / 3h ago / Jun 12" style.
+import { relativeTime as age } from "@/lib/relative-time";
+import { useDateTimePrefs } from "@/lib/datetime-format";
+import { RelativeTime } from "@/components/ui/relative-time";
+import type { Familiar, SessionRow } from "@/lib/types";
+import { FamiliarAvatar } from "@/components/familiar-avatar";
+import { AuthedImage } from "@/components/ui/authed-image";
+import { FamiliarsMemoryView, MemoryFilesList } from "@/components/familiars-memory-view";
+import type { FileMemoryEntry, MemoryFeed } from "@/components/familiars-memory-view";
+import { FamiliarDailyNotes } from "@/components/familiar-daily-notes";
+import { HomeFeed } from "@/components/home/home-feed";
+import { Modal } from "@/components/ui/modal";
+import { EmptyState } from "@/components/ui/empty-state";
+import { Skeleton } from "@/components/ui/skeleton";
+import { Button } from "@/components/ui/button";
+import { FamiliarSummoningCircle } from "@/components/familiar-summoning-circle";
+import {
+  ACTIVITY_DAYS,
+  buildFamiliarCardStats,
+  type FamiliarCardStats,
+  type CovenMemoryEntry,
+} from "@/components/familiars-view-stats";
+import { deriveRenown } from "@/lib/familiar-renown";
+import { compactCount } from "@/lib/profile-card";
+import { useResolvedFamiliars, type ResolvedFamiliar } from "@/lib/familiar-resolve";
+import { SUMMON_FAMILIAR_EVENT, consumeSummonPending } from "@/lib/summon-events";
+import { useFamiliarStudio } from "@/lib/familiar-studio-context";
+import { Popover, PopoverBody, PopoverItem, PopoverSeparator } from "@/components/ui/popover";
+import { SessionTraceOverlay, type TraceTarget } from "@/components/session-trace-overlay";
+
+export function emptyStats(): FamiliarCardStats {
+  return {
+    memoryCount: 0,
+    latestMemory: null,
+    lastSessionAt: null,
+    sessionsTotal: 0,
+    sessionsLast7d: 0,
+    hasActiveSession: false,
+    streakDays: 0,
+    activity: new Array<number>(ACTIVITY_DAYS).fill(0),
+  };
+}
+
+export function FamiliarsEmptyState({
+  onCreate,
+  onOpenOnboarding,
+}: {
+  onCreate: () => void;
+  onOpenOnboarding: () => void;
+}) {
+  return (
+    <EmptyState
+      className="familiars-view__empty mx-auto my-16 max-w-md"
+      icon="ph:sparkle"
+      headline="The circle awaits"
+      subtitle="A familiar is an AI agent with its own identity, memory, and runtime. Summon your first — it can run on this machine, on a remote host over SSH, or bridge an OpenClaw agent you already keep."
+      actions={
+        <div className="flex items-center gap-2">
+          <Button variant="primary" leadingIcon="ph:magic-wand-fill" onClick={onCreate}>
+            Summon a familiar
+          </Button>
+          <Button variant="ghost" onClick={onOpenOnboarding}>
+            Run full setup
+          </Button>
+        </div>
+      }
+    />
+  );
+}
+
+type MemoryStatus = "loading" | "error" | "ready";
+
+type AgentRosterCardProps = {
+  familiar: ResolvedFamiliar;
+  stats: FamiliarCardStats;
+  daemonRunning: boolean;
+  responseNeeded: boolean;
+  memoryStatus: MemoryStatus;
+  onSelect: () => void;
+};
+
+export function FamiliarRosterCard({
+  familiar,
+  stats,
+  daemonRunning,
+  responseNeeded,
+  memoryStatus,
+  onSelect,
+}: AgentRosterCardProps) {
+  const lastSessionLabel = stats.lastSessionAt
+    ? `last session ${age(stats.lastSessionAt)}`
+    : "No sessions yet";
+  const stripTotal = stats.activity.reduce((sum, n) => sum + n, 0);
+  // Renown — earned standing derived from real work (sessions + curated
+  // memories), never a synthetic counter. Tier reads as quiet metadata in the
+  // microlabel voice; the tooltip carries the score and the next rung.
+  const renown = deriveRenown({ sessionsTotal: stats.sessionsTotal, memoryCount: stats.memoryCount });
+  const renownTitle = renown.next
+    ? `${renown.score} renown · ${renown.next.remaining} to ${renown.next.tier.label}`
+    : `${renown.score} renown · top of the ladder`;
+  const streakTitle =
+    stats.streakDays > 0
+      ? `${stats.streakDays} consecutive day${stats.streakDays === 1 ? "" : "s"} with sessions`
+      : "No active streak — a session today starts one";
+  return (
+    // De-boxed card (cave-g2r6) restyled as a terminal stat tile (cave-uw7c):
+    // the wrapper carries the wash + soft hairline so the open button and the
+    // profile/analytics links can sit inside one visual card as sibling
+    // interactive elements (no nested controls). Typography and the stat band
+    // follow the profile share card (src/styles/profile-card.css) — monospace
+    // lowercase labels, hairline-divided tiles, glowing numerals.
+    <div className="familiars-view__card group relative flex h-full flex-col">
+      <button
+        type="button"
+        onClick={onSelect}
+        className="focus-ring flex flex-1 flex-col items-stretch rounded-[inherit] text-left"
+        aria-label={`Open ${familiar.display_name}`}
+      >
+        <span className="flex items-center gap-2.5 px-3 pt-3">
+          <span className="familiars-view__avatar-tile" aria-hidden="true">
+            <FamiliarAvatar familiar={familiar} size="lg" />
+          </span>
+          <span className="min-w-0 flex-1">
+            <span className="block truncate font-mono text-[length:var(--text-base)] font-semibold tracking-wide text-[var(--text-primary)]">
+              {familiar.display_name}
+            </span>
+            <span className="familiars-view__microlabel mt-1 flex min-w-0 items-center gap-1">
+              <span className="truncate">{familiar.role || familiar.harness || familiar.id}</span>
+              <span aria-hidden="true">·</span>
+              <span className="shrink-0 text-[var(--text-secondary)]" title={renownTitle}>
+                {renown.tier.label}
+              </span>
+            </span>
+          </span>
+          <span className="familiars-view__microlabel inline-flex shrink-0 items-center gap-1.5 self-start">
+            <span
+              className={`inline-flex h-1.5 w-1.5 rounded-full ${daemonRunning ? "bg-[var(--accent-presence)] familiars-view__status-dot--live" : "bg-[var(--text-muted)]"}`}
+              aria-hidden="true"
+            />
+            {daemonRunning ? "online" : "offline"}
+          </span>
+        </span>
+
+        {stats.hasActiveSession || responseNeeded ? (
+          <span className="flex flex-wrap items-center gap-1.5 px-3 pt-2">
+            {stats.hasActiveSession ? (
+              <span className="familiars-view__chip text-[var(--accent-presence)]">
+                active session
+              </span>
+            ) : null}
+            {responseNeeded ? (
+              <span className="familiars-view__chip familiars-view__chip--warning">
+                response needed
+              </span>
+            ) : null}
+          </span>
+        ) : null}
+
+        <span className="familiars-view__statband mt-3">
+          <span className="familiars-view__stat">
+            <span className="familiars-view__stat-label">sessions</span>
+            <span className="familiars-view__stat-value">{compactCount(stats.sessionsTotal)}</span>
+          </span>
+          <span className="familiars-view__stat">
+            <span className="familiars-view__stat-label">this week</span>
+            <span className="familiars-view__stat-value">{compactCount(stats.sessionsLast7d)}</span>
+          </span>
+          <span className="familiars-view__stat">
+            <span className="familiars-view__stat-label">memories</span>
+            <span className="familiars-view__stat-value">{compactCount(stats.memoryCount)}</span>
+          </span>
+          <span className="familiars-view__stat" title={streakTitle}>
+            <span className="familiars-view__stat-label">streak</span>
+            <span className="familiars-view__stat-value">
+              {stats.streakDays > 0 ? `${stats.streakDays}d` : "—"}
+            </span>
+          </span>
+        </span>
+
+        <span
+          className="familiars-view__strip"
+          aria-hidden="true"
+          title={`${stripTotal} session${stripTotal === 1 ? "" : "s"} in the last ${ACTIVITY_DAYS} days`}
+        >
+          {stats.activity.map((count, index) => (
+            <i key={index} data-level={activityLevel(count, stats.activity)} />
+          ))}
+        </span>
+
+        <span className="familiars-view__microlabel mt-auto block truncate px-3 pb-2.5 pt-2">
+          {lastSessionLabel}
+        </span>
+      </button>
+
+      {/* Card footer — memory snapshot as a quiet one-liner + the profile and
+          analytics links folded inside the card (they used to float orphaned
+          below the border). Hairline divider, no second box. */}
+      <div className="familiars-view__card-footer flex items-center justify-between gap-2 border-t border-[var(--border-hairline)]/60 px-3 py-2 font-mono text-[length:var(--text-2xs)]">
+        <span className="min-w-0 flex-1 truncate text-[var(--text-muted)]" title={stats.latestMemory?.title}>
+          {memoryStatus === "loading" ? (
+            // Shimmer instead of a "Loading memory…" string — one loading
+            // language across the roster, and no dead-looking text while the
+            // first fetch is cold (cave-5qmm).
+            <span aria-hidden className="block py-0.5">
+              <Skeleton variant="text-sm" width="55%" />
+            </span>
+          ) : memoryStatus === "error" ? (
+            "Memory unavailable"
+          ) : stats.memoryCount === 0 ? (
+            "No memories yet"
+          ) : (
+            <>
+              {stats.memoryCount} memor{stats.memoryCount === 1 ? "y" : "ies"}
+              {stats.latestMemory ? ` · last write ${age(stats.latestMemory.updatedAt)}` : ""}
+            </>
+          )}
+        </span>
+        <Link
+          href={`/dashboard/familiars/${encodeURIComponent(familiar.id)}/profile`}
+          aria-label={`Open profile for ${familiar.display_name}`}
+          className="focus-ring shrink-0 rounded-[var(--radius-sm)] lowercase text-[var(--text-muted)] transition-colors hover:text-[var(--accent-presence)]"
+        >
+          Profile →
+        </Link>
+        <Link
+          href={`/dashboard/familiars/${encodeURIComponent(familiar.id)}/analytics`}
+          aria-label={`Open analytics for ${familiar.display_name}`}
+          className="focus-ring shrink-0 rounded-[var(--radius-sm)] lowercase text-[var(--text-muted)] transition-colors hover:text-[var(--accent-presence)]"
+        >
+          Analytics →
+        </Link>
+      </div>
+    </div>
+  );
+}
+
+/**
+ * Bucket a day's count against the strip max — same 1..4 ramp as the profile
+ * card heatmap (src/lib/profile-card.ts levelFor) so both surfaces agree.
+ */
+function activityLevel(count: number, activity: number[]): 0 | 1 | 2 | 3 | 4 {
+  const max = Math.max(0, ...activity);
+  if (count <= 0 || max <= 0) return 0;
+  return Math.min(4, Math.max(1, Math.ceil((count / max) * 4))) as 1 | 2 | 3 | 4;
+}
+
+// ────────────────────────────────────────────────────────────────────────────
+// FamiliarMemoryOverlay — modal-style full-screen overlay for selected familiar memory
+// ────────────────────────────────────────────────────────────────────────────
+
+type AgentMemoryOverlayProps = {
+  familiars: ResolvedFamiliar[];
+  familiar: ResolvedFamiliar;
+  memoryFeed: MemoryFeed;
+  onClose: () => void;
+  onOpenMemoryFile: (path: string) => void;
+};
+
+export function FamiliarMemoryOverlay({ familiars, familiar, memoryFeed, onClose, onOpenMemoryFile }: AgentMemoryOverlayProps) {
+  const panelRef = useRef<HTMLDivElement>(null);
+  // Trap focus inside the panel + Escape-to-close + restore focus to the opener.
+  useFocusTrap(true, panelRef, { onEscape: onClose });
+
+  return (
+    <div
+      className="familiars-view__overlay fixed inset-0 z-50 grid place-items-center bg-black/40 backdrop-blur-sm"
+      role="dialog"
+      aria-modal="true"
+      aria-label={`Memory for ${familiar.display_name}`}
+      onClick={onClose}
+    >
+      <div
+        ref={panelRef}
+        tabIndex={-1}
+        className="familiars-view__overlay-panel relative flex h-[100dvh] w-full flex-col overflow-hidden border border-[var(--border-hairline)] bg-[var(--bg-base)] shadow-2xl focus:outline-none md:h-[85vh] md:w-[90vw] md:max-w-[1280px] md:rounded-xl"
+        onClick={(e) => e.stopPropagation()}
+      >
+        <button
+          type="button"
+          onClick={onClose}
+          className="focus-ring absolute right-3 top-3 z-10 inline-flex h-7 items-center gap-1 rounded-md border border-[var(--border-hairline)] bg-[var(--bg-raised)] px-2 text-[length:var(--text-xs)] text-[var(--text-secondary)] hover:bg-[var(--bg-raised)]/80"
+          aria-label="Close"
+        >
+          <Icon name="ph:x" width={12} />
+          Close
+        </button>
+        <FamiliarsMemoryView
+          familiars={familiars}
+          activeFamiliar={familiar}
+          lockToFamiliar
+          feed={memoryFeed}
+          onOpenMemoryFile={onOpenMemoryFile}
+        />
+      </div>
+    </div>
+  );
+}
+
+// ────────────────────────────────────────────────────────────────────────────
+// FamiliarDetailRail — thin left avatar column in detail mode
+// ────────────────────────────────────────────────────────────────────────────
+
+type AgentDetailRailProps = {
+  familiars: ResolvedFamiliar[];
+  selectedId: string;
+  onSelect: (id: string) => void;
+  onPreview: (familiar: ResolvedFamiliar) => void;
+  onBack: () => void;
+};
+
+export function FamiliarDetailRail({ familiars, selectedId, onSelect, onPreview, onBack }: AgentDetailRailProps) {
+  return (
+    <nav className="familiars-view__rail flex w-[64px] shrink-0 flex-col items-center gap-2 border-r border-[var(--border-hairline)] bg-[var(--bg-raised)]/20 py-3">
+      <button
+        type="button"
+        onClick={onBack}
+        className="focus-ring familiars-view__rail-back inline-flex h-8 w-8 items-center justify-center rounded-md border border-[var(--border-hairline)] bg-[var(--bg-raised)]/40 text-[var(--text-secondary)] hover:bg-[var(--bg-raised)]"
+        aria-label="Back to roster"
+        title="Back to roster"
+      >
+        <Icon name="ph:arrow-left" width={14} />
+      </button>
+      <div className="mt-1 h-px w-8 bg-[var(--border-hairline)]" aria-hidden="true" />
+      <ul className="flex flex-col items-center gap-1.5">
+        {familiars.map((f) => {
+          const active = f.id === selectedId;
+          return (
+            <li key={f.id}>
+              <button
+                type="button"
+                onClick={() => {
+                  onSelect(f.id);
+                  onPreview(f);
+                }}
+                className={`focus-ring familiars-view__rail-avatar inline-flex h-9 w-9 items-center justify-center overflow-hidden rounded-md border ${
+                  active
+                    ? "border-[var(--accent-presence)] bg-[var(--accent-presence)]/15 text-[var(--accent-presence)]"
+                    : "border-[var(--border-hairline)] bg-[var(--bg-raised)]/40 text-[var(--text-secondary)] hover:bg-[var(--bg-raised)]"
+                }`}
+                title={f.display_name}
+                aria-label={`Preview ${f.display_name}'s avatar`}
+                aria-current={active ? "true" : undefined}
+              >
+                <FamiliarAvatar familiar={f} size="sm" />
+              </button>
+            </li>
+          );
+        })}
+      </ul>
+    </nav>
+  );
+}
+
+// ────────────────────────────────────────────────────────────────────────────
+// FamiliarDetailPanel — right-side panel with Memory / Files / Sessions tabs
+// ────────────────────────────────────────────────────────────────────────────
+
+type DetailTab = "memory" | "daily-notes" | "files" | "sessions" | "feed";
+
+const DETAIL_TABS: { id: DetailTab; label: string }[] = [
+  { id: "memory", label: "Memory" },
+  { id: "daily-notes", label: "Daily Notes" },
+  { id: "files", label: "Files" },
+  { id: "sessions", label: "Sessions" },
+  { id: "feed", label: "Feed" },
+];
+
+type AgentDetailPanelProps = {
+  familiar: ResolvedFamiliar;
+  familiars: ResolvedFamiliar[];
+  sessions: SessionRow[];
+  fileEntries: FileMemoryEntry[];
+  memoryError: string | null;
+  memoryLoaded: boolean;
+  memoryFeed: MemoryFeed;
+  onClose: () => void;
+  onPreview: () => void;
+  onStartChat: () => void;
+  /** Open the summoning circle in Enhancement Rite mode for this familiar. */
+  onEnhance: () => void;
+  onOpenSession: (sessionId: string) => void;
+  onOpenMemoryFile: (path: string) => void;
+  onOpenUrl: (url: string) => void;
+};
+
+// Per-familiar overflow menu on the detail panel header. Remove routes to the
+// Studio lifecycle tab rather than confirming here — the destructive flow
+// (confirm copy, undo toast, tombstone coupling) lives only in
+// familiar-studio-lifecycle-tab.tsx, and this menu is the discoverable
+// entry point the Familiars surface lacked.
+function FamiliarPanelMenu({ familiar }: { familiar: ResolvedFamiliar }) {
+  const { openFamiliarStudio } = useFamiliarStudio();
+  const [open, setOpen] = useState(false);
+  const triggerRef = useRef<HTMLButtonElement | null>(null);
+  return (
+    <>
+      <button
+        ref={triggerRef}
+        type="button"
+        className="focus-ring inline-flex h-7 w-7 items-center justify-center rounded-md border border-[var(--border-hairline)] text-[var(--text-secondary)] hover:bg-[var(--bg-raised)]"
+        aria-label={`${familiar.display_name} options`}
+        aria-haspopup="menu"
+        aria-expanded={open}
+        title="Familiar options"
+        onClick={() => setOpen((v) => !v)}
+      >
+        <Icon name="ph:dots-three-vertical" width={14} aria-hidden />
+      </button>
+      <Popover
+        open={open}
+        onOpenChange={setOpen}
+        anchorRef={triggerRef}
+        placement="bottom-end"
+        minWidth={200}
+        ariaLabel="Familiar options"
+      >
+        <PopoverBody>
+          <PopoverItem
+            icon="ph:pencil-simple"
+            onSelect={() => {
+              setOpen(false);
+              openFamiliarStudio(familiar.id, "identity");
+            }}
+          >
+            Edit in Studio
+          </PopoverItem>
+          <PopoverSeparator />
+          <PopoverItem
+            icon="ph:trash"
+            danger
+            onSelect={() => {
+              setOpen(false);
+              openFamiliarStudio(familiar.id, "lifecycle");
+            }}
+          >
+            Remove familiar…
+          </PopoverItem>
+        </PopoverBody>
+      </Popover>
+    </>
+  );
+}
+
+export function FamiliarDetailPanel({
+  familiar,
+  familiars,
+  sessions,
+  fileEntries,
+  memoryError,
+  memoryLoaded,
+  memoryFeed,
+  onClose,
+  onPreview,
+  onStartChat,
+  onEnhance,
+  onOpenSession,
+  onOpenMemoryFile,
+  onOpenUrl,
+}: AgentDetailPanelProps) {
+  const [tab, setTab] = useState<DetailTab>("memory");
+  // Session trace overlay — the daemon event timeline behind one session.
+  const [traceTarget, setTraceTarget] = useState<TraceTarget | null>(null);
+  const familiarSessions = useMemo(
+    () =>
+      sessions
+        .filter((s) => s.familiarId === familiar.id)
+        .sort((a, b) => (a.updated_at < b.updated_at ? 1 : -1)),
+    [sessions, familiar.id],
+  );
+  const familiarFileEntries = useMemo(
+    () =>
+      fileEntries
+        .filter((entry) => entry.familiarId === familiar.id)
+        .sort((a, b) => (a.modified < b.modified ? 1 : -1)),
+    [fileEntries, familiar.id],
+  );
+
+  return (
+    <section className="familiars-view__panel flex min-h-0 flex-1 flex-col bg-[var(--bg-base)]">
+      <header className="flex items-center justify-between gap-2 border-b border-[var(--border-hairline)] px-4 py-3">
+        <div className="flex items-center gap-2">
+          <button
+            type="button"
+            onClick={onPreview}
+            className="focus-ring inline-flex h-12 w-12 shrink-0 items-center justify-center overflow-hidden rounded-lg border border-[var(--border-hairline)] bg-[var(--bg-raised)]/50 text-[var(--text-primary)] hover:border-[var(--accent-presence)]/60"
+            aria-label={`Enlarge ${familiar.display_name}'s avatar`}
+            title="Enlarge avatar"
+          >
+            <FamiliarAvatar familiar={familiar} size="xl" />
+          </button>
+          <div className="min-w-0">
+            <h2 className="familiars-view__nameplate truncate">
+              {familiar.display_name}
+            </h2>
+            <p className="familiars-view__microlabel mt-1 truncate">
+              {familiar.role || familiar.harness || familiar.id}
+            </p>
+          </div>
+        </div>
+        <div className="flex items-center gap-1.5">
+          <button
+            type="button"
+            onClick={onStartChat}
+            title="Start chat"
+            className="focus-ring inline-flex h-7 items-center gap-1 rounded-md border border-[var(--border-hairline)] bg-[var(--bg-raised)] px-2 text-[length:var(--text-xs)] text-[var(--text-primary)] hover:bg-[var(--bg-raised)]/80"
+          >
+            <Icon name="ph:chat-circle-dots" width={12} />
+            Start
+          </button>
+          <button
+            type="button"
+            onClick={onEnhance}
+            title={`Enhance ${familiar.display_name} in the circle`}
+            className="focus-ring inline-flex h-7 items-center gap-1 rounded-md border border-[var(--border-hairline)] bg-[var(--accent-presence)]/10 px-2 text-[length:var(--text-xs)] text-[var(--accent-presence)] hover:bg-[var(--accent-presence)]/15"
+          >
+            <Icon name="ph:magic-wand-fill" width={12} />
+            Enhance
+          </button>
+          <FamiliarPanelMenu familiar={familiar} />
+          <button
+            type="button"
+            onClick={onClose}
+            className="focus-ring inline-flex h-7 items-center gap-1 rounded-md border border-[var(--border-hairline)] px-2 text-[length:var(--text-xs)] text-[var(--text-secondary)] hover:bg-[var(--bg-raised)]"
+            aria-label="Back to roster"
+          >
+            <Icon name="ph:x" width={12} />
+            Close
+          </button>
+        </div>
+      </header>
+
+      <Tabs
+        items={DETAIL_TABS}
+        value={tab}
+        onChange={setTab}
+        ariaLabel="Familiar details"
+        idPrefix="familiar-detail"
+        className="shrink-0 px-3"
+      />
+
+      <div
+        className="flex min-h-0 flex-1 flex-col overflow-hidden"
+        role="tabpanel"
+        id={`familiar-detail-panel-${tab}`}
+        aria-labelledby={`familiar-detail-tab-${tab}`}
+      >
+        {tab === "memory" ? (
+          <FamiliarsMemoryView
+            familiars={familiars}
+            activeFamiliar={familiar}
+            lockToFamiliar
+            feed={memoryFeed}
+            onOpenMemoryFile={onOpenMemoryFile}
+          />
+        ) : tab === "daily-notes" ? (
+          <FamiliarDailyNotes familiar={familiar} />
+        ) : tab === "files" ? (
+          <div className="flex min-h-0 flex-1 flex-col p-4">
+            <div className="mb-2 flex shrink-0 items-center justify-between">
+              <h3 className="text-[length:var(--text-xs)] font-semibold uppercase tracking-widest text-[var(--text-secondary)]">
+                Memory files
+              </h3>
+              <span className="text-[length:var(--text-2xs)] text-[var(--text-muted)]">
+                {familiarFileEntries.length} total
+              </span>
+            </div>
+            <MemoryFilesList
+              entries={familiarFileEntries}
+              loaded={memoryLoaded}
+              error={memoryError}
+              onOpen={onOpenMemoryFile}
+              className="flex min-h-0 flex-1 flex-col overflow-hidden"
+              listClassName="h-full min-h-0 divide-y divide-[var(--border-hairline)] overflow-y-auto"
+            />
+            <p className="mt-2 text-[length:var(--text-2xs)] text-[var(--text-muted)]">
+              Only files traced to {familiar.display_name} are shown here.
+            </p>
+          </div>
+        ) : tab === "feed" ? (
+          <div className="flex min-h-0 flex-1 flex-col overflow-y-auto p-4">
+            <HomeFeed onOpenUrl={onOpenUrl} />
+          </div>
+        ) : (
+          <div className="h-full overflow-y-auto p-4">
+            <div className="mb-2 flex items-center justify-between">
+              <h3 className="text-[length:var(--text-xs)] font-semibold uppercase tracking-widest text-[var(--text-secondary)]">
+                Sessions
+              </h3>
+              <span className="text-[length:var(--text-2xs)] text-[var(--text-muted)]">
+                {familiarSessions.length} total
+              </span>
+            </div>
+            {familiarSessions.length === 0 ? (
+              <EmptyState
+                compact
+                icon="ph:chats-circle"
+                headline="No sessions for this familiar yet"
+                subtitle="Sessions started with this familiar will show up here."
+              />
+            ) : (
+              <ul className="divide-y divide-[var(--border-hairline)] rounded-lg border border-[var(--border-hairline)] bg-[var(--bg-raised)]/25">
+                {familiarSessions.map((s) => (
+                  <li key={s.id} className="flex items-stretch">
+                    <button
+                      type="button"
+                      onClick={() => onOpenSession(s.id)}
+                      className="focus-ring-inset flex min-w-0 flex-1 items-start gap-2 px-3 py-2 text-left hover:bg-[var(--bg-raised)]"
+                    >
+                      <Icon name="ph:terminal-window" width={13} className="mt-0.5 shrink-0 text-[var(--text-muted)]" />
+                      <span className="min-w-0 flex-1">
+                        <span className="block truncate text-[length:var(--text-sm)] text-[var(--text-primary)]">
+                          {s.title || s.id}
+                        </span>
+                        <span className="mt-0.5 block truncate font-mono text-[length:var(--text-2xs)] text-[var(--text-muted)]">
+                          {s.harness} · {s.status}
+                        </span>
+                      </span>
+                      <RelativeTime iso={s.updated_at} className="shrink-0 text-[length:var(--text-2xs)] text-[var(--text-muted)]" />
+                    </button>
+                    <button
+                      type="button"
+                      onClick={() => setTraceTarget({ id: s.id, title: s.title })}
+                      className="focus-ring-inset flex shrink-0 items-center gap-1 border-l border-[var(--border-hairline)] px-2 text-[length:var(--text-2xs)] text-[var(--text-secondary)] hover:bg-[var(--bg-raised)] hover:text-[var(--text-primary)]"
+                      title="Trace this session's daemon events"
+                      aria-label={`Trace ${s.title || s.id}`}
+                    >
+                      <Icon name="ph:tree-structure" width={12} aria-hidden />
+                      Trace
+                    </button>
+                  </li>
+                ))}
+              </ul>
+            )}
+          </div>
+        )}
+      </div>
+
+      {traceTarget ? (
+        <SessionTraceOverlay target={traceTarget} onClose={() => setTraceTarget(null)} />
+      ) : null}
+    </section>
+  );
+}
+
+// ────────────────────────────────────────────────────────────────────────────
+// FamiliarAvatarPreviewOverlay — enlarged peek at the selected familiar avatar
+// ────────────────────────────────────────────────────────────────────────────
+
+type FamiliarAvatarPreviewOverlayProps = {
+  familiar: ResolvedFamiliar;
+  onClose: () => void;
+};
+
+export function FamiliarAvatarPreviewOverlay({ familiar, onClose }: FamiliarAvatarPreviewOverlayProps) {
+  return (
+    <Modal
+      open
+      onClose={onClose}
+      breadcrumb={["Familiars", familiar.display_name]}
+      ariaLabel={`${familiar.display_name} avatar preview`}
+    >
+      <div className="flex flex-col items-center gap-3 text-center">
+        <div className="grid aspect-square w-full max-w-[320px] place-items-center overflow-hidden rounded-xl border border-[var(--border-hairline)] bg-[var(--bg-base)]">
+          <AuthedImage
+            src={familiar.avatarImage}
+            alt={`${familiar.display_name} avatar`}
+            className="h-full w-full object-cover"
+            fallback={<FamiliarAvatar familiar={familiar} size="xl" />}
+          />
+        </div>
+        <div className="min-w-0">
+          <div className="truncate text-[length:var(--text-md)] font-semibold text-[var(--text-primary)]">
+            {familiar.display_name}
+          </div>
+          <div className="familiars-view__microlabel mt-1 truncate">
+            {familiar.role || familiar.harness || familiar.id}
+          </div>
+        </div>
+      </div>
+    </Modal>
+  );
+}
